changes on changes password

+++ resources/views/admin/occupation/index.blade.php
@@ -1,18 +1,19 @@
 @extends('layouts.app')
 @section('title')@lang('quickadmin.occupation.title')@endsection
 @section('customCss')
-<meta name="csrf-token" content="{{ csrf_token() }}" >
 @endsection
 
 @section('main-content')
 
-
-<div class="right-content">
     <div class="animate__animated animate__fadeInUp">
         <div class="msg-content white-bg radius-50 space-30 d-flex align-items-center">
             <h2 class="mb-md-0">@lang('cruds.occupation.title_singular')</h2>
-            <a href="javascript:void(0)" class="add_btn dash-btn green-bg w-115" id="addOccupationBtn" >+ @lang('global.add') @lang('global.new')</a><div class="d-sm-none w-100"></div>
-            <button class="del_btn dash-btn red-bg w-115 me-md-1" id="deleteAllOccupation">@lang('global.delete')</button>
+            @can('occupation_create')
+                <a href="javascript:void(0)" class="add_btn dash-btn green-bg w-115" id="addOccupationBtn" >+ @lang('global.add') @lang('global.new')</a><div class="d-sm-none w-100"></div>
+            @endcan
+            @can('occupation_delete')
+                <button class="del_btn dash-btn red-bg w-115 me-md-1" id="deleteAllOccupation">@lang('global.delete')</button>
+            @endcan
         </div>
         <div class="c-admin position-relative">
             <div class="table-responsive">
@@ -20,8 +21,6 @@
             </div>
         </div>
     </div>
-</div>
-
 
 @endsection
 @section('customJS')
@@ -44,6 +43,11 @@
                         $('.popup_render_div').html(response.htmlView);
                         $('#addOccupationModal').modal('show');
                     }
+                },
+                error: function (response) {
+                    if(response.responseJSON.error_type == 'something_error'){
+                        toasterAlert('error',response.responseJSON.error);
+                    } 
                 }
             })
         });
@@ -66,17 +70,17 @@
                 success: function (response) {                
                     if(response.success) {
                         $('#addOccupationModal').modal('hide');
-                        fireSuccessSwal('Success',response.message);
+                        toasterAlert('success',response.message);
                         $('#occupation-table').DataTable().ajax.reload(null, false);
                     }
                 },
                 error: function (response) {
-                    if(response.error_type == 'something_error'){
-                        fireErrorSwal('Error',response.message);
+                    if(response.responseJSON.error_type == 'something_error'){
+                        toasterAlert('error',response.responseJSON.error);
                     } else {                    
                         var errorLabelTitle = '';
                         $.each(response.responseJSON.errors, function (key, item) {
-                            errorLabelTitle = '<span class="validation-error-block">'+item+'</sapn>';
+                            errorLabelTitle = '<span class="validation-error-block">'+item[0]+'</sapn>';
                             if(key == 'resume' || key == 'gender'){
                                 $('#'+key).html(errorLabelTitle);
                             }
@@ -109,6 +113,11 @@
                         $('.popup_render_div').html(response.htmlView);
                         $('#editOccupationModal').modal('show');
                     }
+                },
+                error: function (response) {
+                    if(response.responseJSON.error_type == 'something_error'){
+                        toasterAlert('error',response.responseJSON.error);
+                    } 
                 }
             });
         });
@@ -132,18 +141,18 @@
                 success: function (response) {
                     if(response.success) {
                         $('#editOccupationModal').modal('hide');
-                        fireSuccessSwal('Success',response.message);
+                        toasterAlert('success',response.message);
                         $('#occupation-table').DataTable().ajax.reload(null, false);
                     }
                 },
                 error: function (response) {
                     $(".submitBtn").attr('disabled', false);
-                    if(response.error_type == 'something_error'){
-                        fireErrorSwal('Error',response.message);
+                    if(response.responseJSON.error_type.error_type == 'something_error'){
+                        toasterAlert('error',response.responseJSON.error_type.error);
                     } else {                    
                         var errorLabelTitle = '';
                         $.each(response.responseJSON.errors, function (key, item) {
-                            errorLabelTitle = '<span class="validation-error-block">'+item+'</sapn>';
+                            errorLabelTitle = '<span class="validation-error-block">'+item[0]+'</sapn>';
                             if(key == 'resume' || key == 'gender'){
                                 $('#'+key).html(errorLabelTitle);
                             }
@@ -180,15 +189,15 @@
                         data: { _token: "{{ csrf_token() }}" },
                         success: function (response) {
                             if(response.success) {
-                                fireSuccessSwal('Success',response.message);
+                                toasterAlert('success',response.message);
                                 $('#occupation-table').DataTable().ajax.reload(null, false);
                             }
                             else {
-                                fireErrorSwal('Error',response.message);
+                                toasterAlert('error',response.message);
                             }
                         },
                         error: function(res){
-                            fireErrorSwal('Error',res.message);
+                            toasterAlert('error',res.responseJSON.error);
                         }
                     });
                 }
@@ -228,15 +237,15 @@
                         dataType: 'json',
                         success: function (response) {
                             if(response.success) {
-                                fireSuccessSwal('Success',response.message);
+                                toasterAlert('success',response.message);
                                 $('#occupation-table').DataTable().ajax.reload(null, false);
                             }
                             else {
-                                fireErrorSwal('Error',response.message);
+                                toasterAlert('error',response.message);
                             }
                         },
                         error: function(res){
-                            fireErrorSwal('Error',res.message);
+                            toasterAlert('error',res.responseJSON.error);
                         }
                     })
                 }
